requirements.txt
# # Core Dependencies
# pydantic>=2.5.0
# pyyaml>=6.0.1
# python-dotenv>=1.0.0
# jinja2>=3.1.2
# typing-extensions>=4.8.0

# # Logging and Monitoring
# structlog>=23.2.0
# colorama>=0.4.6

# # Audio Processing
# librosa>=0.10.1
# soundfile>=0.12.1
# pydub>=0.25.1
# scipy>=1.11.4
# numpy>=1.24.0

# # Speech Recognition Libraries
# speech-recognition>=3.10.0
# whisper-openai>=20231117
# azure-cognitiveservices-speech>=1.34.0
# google-cloud-speech>=2.21.0

# # Text-to-Speech Libraries
# azure-cognitiveservices-speech>=1.34.0
# google-cloud-texttospeech>=2.16.0
# openai>=1.3.0
# gtts>=2.4.0

# # Web and API
# requests>=2.31.0
# aiohttp>=3.9.0
# httpx>=0.25.0

# # Data Processing
# pandas>=2.1.0
# matplotlib>=3.8.0
# seaborn>=0.13.0

# # Testing
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# pytest-mock>=3.12.0
# pytest-cov>=4.1.0
# pytest-html>=4.1.0

# # Audio Similarity and Analysis
# resemblyzer>=0.1.1
# pesq>=0.0.4
# pypesq>=1.2.4

# # Evaluation Metrics
# jiwer>=3.0.0
# rapidfuzz>=3.5.0

# # Utilities
# click>=8.1.0
# rich>=13.7.0
# tqdm>=4.66.0



# Core Dependencies
pydantic>=2.5.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
jinja2>=3.1.2
typing-extensions>=4.8.0

# Logging and Monitoring
structlog>=23.2.0
colorama>=0.4.6

# Audio Processing
librosa>=0.10.1
soundfile>=0.12.1
pydub>=0.25.1
scipy>=1.11.4
numpy>=1.24.0

# Speech Recognition Libraries
SpeechRecognition>=3.10.0
openai-whisper>=20231117
azure-cognitiveservices-speech>=1.34.0
google-cloud-speech>=2.21.0

# Text-to-Speech Libraries
azure-cognitiveservices-speech>=1.34.0
google-cloud-texttospeech>=2.16.0
openai>=1.3.0
gtts>=2.4.0

# Web and API
requests>=2.31.0
aiohttp>=3.9.0
httpx>=0.25.0

# Data Processing
pandas>=2.1.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0
pytest-html>=4.1.0

# Audio Similarity and Analysis
Resemblyzer>=0.1.1
pesq>=0.0.4
pypesq>=1.2.4

# Evaluation Metrics
jiwer>=3.0.0
rapidfuzz>=3.5.0

# Utilities
click>=8.1.0
rich>=13.7.0
tqdm>=4.66.0